refactor: split timeline motion components
Fixes #46
