+++ packages/editor/src/timeline/keyframeValue.ts
@@ -1,0 +1,4 @@
+export function parseKeyframeValue(value: string): unknown {
+  const numeric = Number(value);
+  return value.trim() !== '' && Number.isFinite(numeric) ? numeric : value;
+}
